--- a/scripts/benchmarks/setup.sh
+++ b/scripts/benchmarks/setup.sh
@@ -4,4 +4,4 @@
 # Setup the CPython repository to enable benchmarking.
 ###
 
-git clone --branch 3.10 https://github.com/python/cpython.git crates/ruff/resources/test/cpython
+git clone --branch 3.10 https://github.com/python/cpython.git crates/ruff/resources/test/cpython --filter=blob:none