```console
# Before
scripts/benchmarks/setup.sh  72,98s user 9,59s system 11% cpu 11:31,73 total
# After
scripts/benchmarks/setup.sh  10,70s user 2,77s system 6% cpu 3:33,95 total
```