@konstin 2023-04-30
thanks, that's a huge speedup!

i have to try this in the ecosystem ci docker too, cloning is currently the bottleneck there