@JonathanPlasse 2023-04-30
The solution used for the ecosystem is much faster than `--filter=blob:none`.