@JonathanPlasse 2023-04-30
> do you mean the settings below? i'm unfortunately not familiar enough with the internals of git to compare them.
> 
> https://github.com/charliermarsh/ruff/blob/a32617911a36998d2ad88ee7bafcd7c9e01720f1/scripts/check_ecosystem.py#L43-L54

Yes, that is what I checked.